extract project preview for clarity
